backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==2.2.4
yfinance==0.2.55
scikit-learn==1.6.1
python-dotenv==1.0.1
